Add initial upower backend
Removes any libhal and bme code.

There's quite a bit of tweaking to be done, at least regarding:

* charging state
* mAh calculation (based on voltage_now...)

The applet itself accurately represents what UPower presents to us, but
UPower itself seems to be quite buggy regarding the values it reports,
it will think the device is charging while /sys reports the device is
discharging. That will have to fixed in UPower.

Or we can write a /sys backend.