Remove vendor matching when searching for battery
It was arbitrary and immediately breaks on the Droid4.

The reason it was added is because the N900 exposes two "batteries", and
I was trying to be clever about finding the right one.

Perhaps we just need a whitelist of "preferred" batteries. And if none
of those are found, we fall back to the normal probe.