batmon: cleanup blacklist
Blacklisting of unneeded drivers for the N900 is now handled by a modprobe
blocklist.

Additionally, the bq2415x driver was fixed so upower correctly sees it as
a charger instead of a battery, thus it does not need to be blacklisted
anymore.

The Droid 4 blacklist appears to have been added in error.